requirements.txt
absl-py==0.8.1
alabaster==0.7.8
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.3.1
astropy==3.0
attrs==17.4.0
Babel==2.4.0
bleach==2.1.2
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
cloudpickle==0.5.2
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
Cython==0.26.1
decorator==4.1.2
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
emcee==2.2.1
entrypoints==0.2.3.post1
fail2ban==0.10.2
future==0.18.1
gast==0.3.2
google-pasta==0.1.7
gpg==1.10.0
grpcio==1.24.1
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
imagesize==0.7.1
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.11.1
Jinja2==2.10
joeynmt==0.0.1
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
logilab-common==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.3
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
nose==1.3.7
nose-parameterized==0.3.4
numpy==1.17.3
numpydoc==0.7.0
oauth==1.0.1
olefile==0.45.1
PAM==0.4.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.0
pluggy==0.6.0
ply==3.11
portalocker==1.5.1
prompt-toolkit==1.0.15
protobuf==3.10.0
psutil==5.4.2
py==1.5.2
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pyds9==1.8.1
pyfits==3.4
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==2.4.2
pymacaroons==0.13.0
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyparsing==2.4.2
pyqtgraph==0.10.0
pyRFC3339==1.0
pytest==3.3.2
pytest-arraydiff==0.2
pytest-astropy==0.2.1
pytest-doctestplus==0.1.2
pytest-openfiles==0.2.0
pytest-remotedata==0.2.0
python-apt==1.6.5+ubuntu0.4
python-dateutil==2.8.0
python-debian==0.1.32
python-magic==0.4.16
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
roman==2.0.0
rope==0.10.5
sacrebleu==1.4.2
scipy==1.3.1
scour==0.36
screen-resolution-extra==0.0.0
seaborn==0.9.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.12.0
Sphinx==1.6.7
spyder==3.2.6
ssh-import-id==5.7
subword-nmt==0.3.6
sympy==1.1.1
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testpath==0.3.1
torch==1.3.0
torchtext==0.4.0
tornado==4.5.3
tqdm==4.36.1
traitlets==4.3.2
typed-ast==1.4.0
typing==3.7.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.25.6
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2

# ISL dataset processing dependencies
sentencepiece>=0.1.96
mediapipe>=0.8.10
opencv-python>=4.5.0
openpyxl>=3.0.0
kaggle>=1.5.12
kagglehub>=0.1.0
fairseq>=0.10.0
onnx>=1.10.0
onnxruntime>=1.10.0
tensorflow>=2.6.0
tensorflow-lite>=2.6.0
